+++ src/Type/Events.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Message\Mothership\User\Type;
+
+/**
+ * Class Events
+ * @package Message\Mothership\User\Type
+ *
+ * @author  Thomas Marchant <thomas@mothership.ec>
+ */
+class Events
+{
+	const PROFILE_UPDATE = 'ms.user.profile.update';
+}