+++ src/Type/UserTypeInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Message\Mothership\User\Type;
+
+use Message\Cog\Field;
+
+/**
+ * Interface UserTypeInterface
+ * @package Message\Mothership\User\Type
+ *
+ * @author  Thomas Marchant <thomas@mothership.ec>
+ */
+interface UserTypeInterface extends Field\ContentTypeInterface
+{
+
+}